[Behat] Extract Shipping transformer
