+++ src/Sylius/Behat/Context/Transform/ShippingContext.php
@@ -1,0 +1,49 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Behat\Context\Transform;
+
+use Behat\Behat\Context\Context;
+use Sylius\Component\Resource\Repository\RepositoryInterface;
+
+/**
+ * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
+ */
+final class ShippingContext implements Context
+{
+    /**
+     * @var RepositoryInterface
+     */
+    private $shippingMethodRepository;
+
+    /**
+     * @param RepositoryInterface $shippingMethodRepository
+     */
+    public function __construct(RepositoryInterface $shippingMethodRepository)
+    {
+        $this->shippingMethodRepository = $shippingMethodRepository;
+    }
+
+    /**
+     * @Transform :shippingMethodName shipping method
+     * @Transform shipping method :shippingMethodName
+     * @Transform /^"([^"]*)" shipping method$/
+     */
+    public function getShippingMethodByName($shippingMethodName)
+    {
+        $shippingMethod = $this->shippingMethodRepository->findOneBy(['name' => $shippingMethodName]);
+        if (null === $shippingMethod) {
+            throw new \Exception('Shipping method with name "'.$shippingMethodName.'" does not exist');
+        }
+
+        return $shippingMethod;
+    }
+}
